+++ laktory/models/resources/databricks/notebook.py
@@ -1,11 +1,11 @@
 import os
 from pathlib import Path
-from typing import Any
 from typing import Literal
 from typing import Union
 
+from pydantic import AliasChoices
 from pydantic import Field
-from pydantic import model_validator
+from pydantic import computed_field
 
 from laktory import settings
 from laktory.models.basemodel import BaseModel
@@ -43,18 +43,16 @@ class Notebook(BaseModel, PulumiResource, TerraformResource):
 
     notebook = models.resources.databricks.Notebook(
         source="./notebooks/dlt/dlt_laktory_pl.py",
-        rootpath="/src/",
     )
     print(notebook.path)
-    # > /src/dlt_laktory_pl.py
+    # > /.laktory/dlt_laktory_pl.py
 
     notebook = models.resources.databricks.Notebook(
         source="./notebooks/dlt/dlt_laktory_pl.py",
-        rootpath="/src/",
         dirpath="notebooks/dlt/",
     )
     print(notebook.path)
-    # > /src/notebooks/dlt/dlt_laktory_pl.py
+    # > /.laktory/notebooks/dlt/dlt_laktory_pl.py
     ```
     """
 
@@ -73,35 +71,21 @@ class Notebook(BaseModel, PulumiResource, TerraformResource):
         exclude=True,
         description="Specifications for looking up existing resource. Other attributes will be ignored.",
     )
-    path: str = Field(
+    path_: str = Field(
         None,
         description="Workspace filepath for the notebook. Overwrite `rootpath` and `dirpath`.",
-    )
-    rootpath: str = Field(
-        None,
-        description="""
-    Root directory to which all notebooks are deployed to. Can also be configured by settings 
-    LAKTORY_WORKSPACE_LAKTORY_ROOT environment variable. Default is `/.laktory/`. Used only if `path` is not specified.
-    """,
+        validation_alias=AliasChoices("path", "path_"),
+        exclude=True,
     )
     source: str = Field(
         ..., description="Path to notebook in source code format on local filesystem."
     )
 
+    @computed_field(description="path")
     @property
-    def filename(self) -> str:
-        """Notebook file name"""
-        return os.path.basename(self.source)
-
-    @model_validator(mode="after")
-    def set_paths(self) -> Any:
-        # Path set
-        if self.path:
-            return self
-
-        # root
-        if self.rootpath is None:
-            self.rootpath = settings.workspace_laktory_root
+    def path(self) -> str | None:
+        if self.path_:
+            return self.path_
 
         # dir
         if self.dirpath is None:
@@ -109,11 +93,13 @@ class Notebook(BaseModel, PulumiResource, TerraformResource):
         if self.dirpath.startswith("/"):
             self.dirpath = self.dirpath[1:]
 
-        # path
-        _path = Path(self.rootpath) / self.dirpath / self.filename
-        self.path = _path.as_posix()
+        path = Path(settings.workspace_laktory_root) / self.dirpath / self.filename
+        return path.as_posix()
 
-        return self
+    @property
+    def filename(self) -> str:
+        """Notebook file name"""
+        return os.path.basename(self.source)
 
     # ----------------------------------------------------------------------- #
     # Resource Properties                                                     #
@@ -160,7 +146,7 @@ class Notebook(BaseModel, PulumiResource, TerraformResource):
 
     @property
     def pulumi_excludes(self) -> Union[list[str], dict[str, bool]]:
-        return ["access_controls", "dirpath", "rootpath"]
+        return ["access_controls", "dirpath"]
 
     # ----------------------------------------------------------------------- #
     # Terraform Properties                                                    #