+++ laktory/models/resources/databricks/workspacefile.py
@@ -1,12 +1,10 @@
 import os
 from pathlib import Path
-from typing import Any
 from typing import Union
 
 from pydantic import AliasChoices
 from pydantic import Field
 from pydantic import computed_field
-from pydantic import model_validator
 
 from laktory import settings
 from laktory.models.basemodel import BaseModel
@@ -33,18 +31,16 @@ class WorkspaceFile(BaseModel, PulumiResource, TerraformResource):
 
     file = models.resources.databricks.WorkspaceFile(
         source="./notebooks/dlt/dlt_laktory_pl.py",
-        rootpath="/src/",
     )
     print(file.path)
-    # > /src/dlt_laktory_pl.py
+    # > /.laktory/dlt_laktory_pl.py
 
     file = models.resources.databricks.WorkspaceFile(
         source="./notebooks/dlt/dlt_laktory_pl.py",
-        rootpath="/src/",
         dirpath="notebooks/dlt/",
     )
     print(file.path)
-    # > /src/notebooks/dlt/dlt_laktory_pl.py
+    # > /.laktory/notebooks/dlt/dlt_laktory_pl.py
     ```
     """
 
@@ -53,16 +49,11 @@ class WorkspaceFile(BaseModel, PulumiResource, TerraformResource):
         None,
         description="Workspace directory inside rootpath in which the workspace file is deployed. Used only if `path` is not specified.",
     )
-    path: str = Field(
+    path_: str = Field(
         None,
-        description="Workspace filepath for the file. Overwrite `rootpath` and `dirpath`.",
-    )
-    rootpath: str = Field(
-        None,
-        description="""
-    Root directory to which all workspace files are deployed to. Can also be configured by settings 
-    LAKTORY_WORKSPACE_LAKTORY_ROOT environment variable. Default is `/.laktory/`. Used only if `path` is not specified.
-    """,
+        description="Workspace filepath for the file. Overwrite `dirpath`.",
+        validation_alias=AliasChoices("path", "path_"),
+        exclude=True,
     )
     source_: str = Field(
         None,
@@ -84,28 +75,14 @@ class WorkspaceFile(BaseModel, PulumiResource, TerraformResource):
     def source(self) -> str:
         return self.source_
 
-    @classmethod
-    def lookup_defaults(cls) -> dict:
-        return {"path": ""}
-
+    @computed_field(description="path")
     @property
-    def filename(self) -> str:
-        """File filename"""
-        if self.source:
-            return os.path.basename(self.source)
-
-    @model_validator(mode="after")
-    def set_paths(self) -> Any:
-        # Path set
-        if self.path:
-            return self
+    def path(self) -> str | None:
+        if self.path_:
+            return self.path_
 
         if not self.source_:
-            return self
-
-        # root
-        if self.rootpath is None:
-            self.rootpath = settings.workspace_laktory_root
+            return None
 
         # dir
         if self.dirpath is None:
@@ -113,11 +90,18 @@ class WorkspaceFile(BaseModel, PulumiResource, TerraformResource):
         if self.dirpath.startswith("/"):
             self.dirpath = self.dirpath[1:]
 
-        # path
-        _path = Path(self.rootpath) / self.dirpath / self.filename
-        self.path = _path.as_posix()
+        path = Path(settings.workspace_laktory_root) / self.dirpath / self.filename
+        return path.as_posix()
 
-        return self
+    @classmethod
+    def lookup_defaults(cls) -> dict:
+        return {"path": ""}
+
+    @property
+    def filename(self) -> str | None:
+        """File filename"""
+        if self.source:
+            return os.path.basename(self.source)
 
     # ----------------------------------------------------------------------- #
     # Resource Properties                                                     #
@@ -160,7 +144,7 @@ class WorkspaceFile(BaseModel, PulumiResource, TerraformResource):
 
     @property
     def pulumi_excludes(self) -> Union[list[str], dict[str, bool]]:
-        return ["access_controls", "dirpath", "rootpath"]
+        return ["access_controls", "dirpath"]
 
     # ----------------------------------------------------------------------- #
     # Terraform Properties                                                    #