+++ tests/resources/test_workspacefile.py
@@ -4,7 +4,6 @@ from laktory.models.resources.databricks import WorkspaceFile
 def get_workspace_file():
     return WorkspaceFile(
         source="./test_workspacefile.py",
-        rootpath="/",
         dirpath="/init_scripts/",
         access_controls=[
             {"permission_level": "CAN_READ", "group_name": "account users"}
@@ -15,11 +14,11 @@ def get_workspace_file():
 def test_workspace_file():
     workspace_file = get_workspace_file()
     assert workspace_file.filename == "test_workspacefile.py"
-    assert workspace_file.path == "/init_scripts/test_workspacefile.py"
-    assert workspace_file.resource_key == "init_scripts-test_workspacefile-py"
+    assert workspace_file.path == "/.laktory/init_scripts/test_workspacefile.py"
+    assert workspace_file.resource_key == "laktory-init_scripts-test_workspacefile-py"
     assert (
         workspace_file.resource_name
-        == "workspace-file-init_scripts-test_workspacefile-py"
+        == "workspace-file-laktory-init_scripts-test_workspacefile-py"
     )
 
     assert workspace_file.access_controls[0].permission_level == "CAN_READ"